# docs/acceptance/openapi-before.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Exam AI Assistant API",
    "description": "Backend service for short-term exam review workflows.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/exams": {
      "post": {
        "tags": [
          "exams"
        ],
        "summary": "Generate Exam",
        "operationId": "generate_exam_exams_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExamGenerateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "exams"
        ],
        "summary": "List Exams",
        "operationId": "list_exams_exams_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "subject_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Subject Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/exams/generations": {
      "post": {
        "tags": [
          "exams"
        ],
        "summary": "Queue Exam Generation",
        "operationId": "queue_exam_generation_exams_generations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExamGenerateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/exams/{exam_id}": {
      "get": {
        "tags": [
          "exams"
        ],
        "summary": "Get Exam",
        "operationId": "get_exam_exams__exam_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "exam_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Exam Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/exams/{exam_id}/export": {
      "get": {
        "tags": [
          "exams"
        ],
        "summary": "Export Exam",
        "operationId": "export_exam_exams__exam_id__export_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "exam_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Exam Id"
            }
          },
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "pdf",
              "title": "Format"
            }
          },
          {
            "name": "includeAnswers",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Includeanswers"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/workbench/exams/{subject_id}": {
      "get": {
        "tags": [
          "workbench-placeholder"
        ],
        "summary": "Exams",
        "operationId": "exams_api_workbench_exams__subject_id__get",
        "parameters": [
          {
            "name": "subject_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Subject Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/workbench/exams": {
      "post": {
        "tags": [
          "workbench-placeholder"
        ],
        "summary": "Create Exam",
        "operationId": "create_exam_api_workbench_exams_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExamCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/exams": {
      "post": {
        "tags": [
          "exams"
        ],
        "summary": "Generate Exam",
        "operationId": "generate_exam_api_v1_exams_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExamGenerateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "exams"
        ],
        "summary": "List Exams",
        "operationId": "list_exams_api_v1_exams_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "subject_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Subject Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/exams/generations": {
      "post": {
        "tags": [
          "exams"
        ],
        "summary": "Queue Exam Generation",
        "operationId": "queue_exam_generation_api_v1_exams_generations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExamGenerateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/exams/{exam_id}": {
      "get": {
        "tags": [
          "exams"
        ],
        "summary": "Get Exam",
        "operationId": "get_exam_api_v1_exams__exam_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "exam_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Exam Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/exams/{exam_id}/export": {
      "get": {
        "tags": [
          "exams"
        ],
        "summary": "Export Exam",
        "operationId": "export_exam_api_v1_exams__exam_id__export_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "exam_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Exam Id"
            }
          },
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "pdf",
              "title": "Format"
            }
          },
          {
            "name": "includeAnswers",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Includeanswers"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ExamCreate": {
        "properties": {
          "subject_id": {
            "type": "string",
            "title": "Subject Id"
          },
          "question_count": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Question Count",
            "default": 10
          },
          "difficulty": {
            "type": "string",
            "title": "Difficulty",
            "default": "medium"
          }
        },
        "type": "object",
        "required": [
          "subject_id"
        ],
        "title": "ExamCreate"
      },
      "ExamGenerateRequest": {
        "properties": {
          "subjectId": {
            "type": "string",
            "title": "Subjectid"
          },
          "title": {
            "type": "string",
            "maxLength": 300,
            "minLength": 1,
            "title": "Title",
            "default": "期末模拟试卷"
          },
          "durationMinutes": {
            "type": "integer",
            "maximum": 1440.0,
            "minimum": 1.0,
            "title": "Durationminutes",
            "default": 90
          },
          "difficulty": {
            "type": "string",
            "enum": [
              "easy",
              "medium",
              "hard"
            ],
            "title": "Difficulty",
            "default": "medium"
          },
          "questionTypes": {
            "items": {
              "$ref": "#/components/schemas/QuestionTypeSpec"
            },
            "type": "array",
            "title": "Questiontypes"
          },
          "scope": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ],
            "title": "Scope"
          },
          "materialIds": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Materialids"
          },
          "externalRatio": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Externalratio",
            "default": 0
          },
          "knowledgePolicy": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "course_only",
                  "course_first",
                  "expanded"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Knowledgepolicy"
          },
          "avoidSeen": {
            "type": "boolean",
            "title": "Avoidseen",
            "default": true
          },
          "assessmentType": {
            "type": "string",
            "enum": [
              "practice",
              "stage",
              "mock"
            ],
            "title": "Assessmenttype",
            "default": "mock"
          }
        },
        "type": "object",
        "required": [
          "subjectId",
          "questionTypes"
        ],
        "title": "ExamGenerateRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "QuestionTypeSpec": {
        "properties": {
          "questionType": {
            "type": "string",
            "enum": [
              "single_choice",
              "multiple_choice",
              "true_false",
              "fill_blank",
              "short_answer",
              "calculation",
              "essay"
            ],
            "title": "Questiontype"
          },
          "count": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Count"
          },
          "pointsEach": {
            "type": "number",
            "maximum": 100.0,
            "exclusiveMinimum": 0.0,
            "title": "Pointseach"
          }
        },
        "type": "object",
        "required": [
          "questionType",
          "count",
          "pointsEach"
        ],
        "title": "QuestionTypeSpec"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}